+++ packages/docstring/index.ts
@@ -1,1 +1,2 @@
 export * from "./lib";
+export * from "./lib/type";
