--- packages/docstring/lib/field/type.ts
@@ -1,8 +1,0 @@
-import { Option } from "@code-ui/type";
-export interface IField<T = string> {
-  tag?: "@";
-  name: string;
-  enabled?: boolean;
-  template?: string;
-  options?: Option<T>[];
-}
