+++ packages/docstring/lib/index.tsx
@@ -1,9 +1,15 @@
 import Field from "./field";
 import React from "react";
 import { CodeComent } from "@code-ui/token";
-import { CodeLikeViewProps, _language_config } from "./type";
+import {
+  CodeLikeViewProps,
+  LanguageType,
+  _DEFAULT_DART,
+  _DEFAULT_JS_STYLE,
+} from "./type";
 import { Line, CodeWrapper, Monokai, Link } from "./style";
 import { ThemeProvider } from "@emotion/react";
+import { LanguageConfig } from "@code-ui/type";
 
 export const Docstring = (props: CodeLikeViewProps) => {
   const expandableConfig = props?.expandableConfig;
@@ -15,6 +21,22 @@ export const Docstring = (props: CodeLikeViewProps) => {
   // const shouldhide = () => {
   //   return expandableConfig.hidable && lines > expandableConfig.lines;
   // };
+
+  function _language_config(type: LanguageType): LanguageConfig {
+    if (typeof type == "string") {
+      if (type === "js") {
+        return _DEFAULT_JS_STYLE;
+      } else if (type === "dart") {
+        return _DEFAULT_DART;
+      }
+      //  TODO: add python preset
+      else {
+        throw `${type} is not a valid preset`;
+      }
+    } else {
+      return type as LanguageConfig;
+    }
+  }
 
   return (
     <ThemeProvider theme={Monokai}>
